refactor(unity-bootstrap-theme): refactor GA code and header scroll
