+++ packages/unity-bootstrap-theme/src/js/googleAnalytics.js
@@ -4,9 +4,56 @@ const googleAnalytics = () => {
     if (dataLayer) dataLayer.push(event);
   };
 
-  // Clicks events
-  const elements = document.querySelectorAll('[data-ga]');
-  elements.forEach((element) =>
+  // Header Clicks events
+  document.querySelectorAll('[data-ga-header]')?.forEach((element) =>
+    element.addEventListener('click', () => {
+      const dropdown = element.getAttribute('aria-expanded');
+
+      const event = element.getAttribute('data-ga-header-event');
+      const action = dropdown
+        ? dropdown === 'false'
+          ? 'open'
+          : 'close'
+        : element.getAttribute('data-ga-header-action');
+      const type = element.getAttribute('data-ga-header-type');
+      const section = element.getAttribute('data-ga-header-section');
+      const text = element.getAttribute('data-ga-header').toLowerCase();
+      const component = element.getAttribute('data-ga-header-component');
+
+      pushGAEvent({
+        name: 'onclick',
+        event: event ? event.toLowerCase() : 'link',
+        action: action ? action.toLowerCase() : 'click',
+        type: type ? type.toLowerCase() : 'internal link',
+        section: section ? section.toLowerCase() : 'main navbar',
+        region: 'navbar',
+        text,
+        ...(component && {
+          component: component.toLowerCase(),
+        }),
+      });
+    })
+  );
+
+  // Header Search change events
+  document.querySelectorAll('[data-ga-input-header-event]')?.forEach((element) =>
+    element.addEventListener('change', (e) => {
+      const event = element
+        .getAttribute('data-ga-input-header-event')
+        .toLowerCase();
+      const action = 'type';
+      const type = 'main search';
+      const section = 'topbar';
+      const region = 'navbar';
+      const name = 'onenter';
+      const text = e.target.value;
+      pushGAEvent({ name, event, action, type, section, region, text });
+    })
+  );
+
+
+  // General Clicks events
+  document.querySelectorAll('[data-ga]')?.forEach((element) =>
     element.addEventListener('click', () => {
       const dropdown = element.getAttribute('aria-expanded');
 
@@ -35,9 +82,10 @@ const googleAnalytics = () => {
       });
     })
   );
+
+
   // Input change events
-  const inputElements = document.querySelectorAll('[data-ga-input]');
-  inputElements.forEach((element) =>
+  document.querySelectorAll('[data-ga-input]')?.forEach((element) =>
     element.addEventListener('change', (e) => {
       const name = element.getAttribute('data-ga-input-name');
       const event = element.getAttribute('data-ga-input-event');
