+++ packages/unity-bootstrap-theme/stories/organisms/global-header/global-header.js
@@ -1,0 +1,12 @@
+export const initGlobalHeader = () => {
+  // Scroll state
+  const handleWindowScroll = () => {
+    const headerEl = document.getElementById('asu-header');
+    const curPos = window.scrollY;
+    curPos > headerEl?.getBoundingClientRect().top
+      ? headerEl?.classList.add('scrolled')
+      : headerEl?.classList.remove('scrolled');
+  };
+
+  window.addEventListener('scroll', handleWindowScroll);
+};
